+++ src/core/server/extensions/playerFunctions/select.ts
@@ -1,4 +1,5 @@
 import * as alt from 'alt-server';
+import { ATHENA_EVENTS_PLAYER } from '../../enums/athena';
 import { SYSTEM_EVENTS } from '../../../shared/enums/system';
 import { Character } from '../../../shared/interfaces/Character';
 import { DEFAULT_CONFIG } from '../../athena/main';
@@ -78,6 +79,7 @@ async function selectCharacter(p: alt.Player, characterData: Partial<Character>)
         BlipController.populateGlobalBlips(p);
         MarkerController.populateGlobalMarkers(p);
         alt.emit(SYSTEM_EVENTS.VOICE_ADD, p);
+        alt.emit(ATHENA_EVENTS_PLAYER.SELECTED_CHARACTER, p);
     }, 500);
 
     // Delete unused data from the Player.