+++ src/core/server/systems/tick.ts
@@ -4,6 +4,7 @@ import { DEFAULT_CONFIG } from '../athena/main';
 import { playerFuncs } from '../extensions/Player';
 import { vehicleFuncs } from '../extensions/Vehicle';
 import { InventoryController } from '../views/inventory';
+import './athenaEvent';
 import './blip';
 import './interaction';
 import './toolbar';